float temperature

+++ IDP2024F/IDP2024F_10.28.2024/IDP2024F_10.28.2024.ino
@@ -20,7 +20,7 @@
 // General variables
 float outVoltage;
 float batteryVoltage;
-int temperature;
+float temperature;
 
 float dutyCycle = IDEAL_DUTY_CYCLE;
